% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFormat2018.R
\name{applyFormat2018}
\alias{applyFormat2018}
\title{Apply the 2018 format to database}
\usage{
applyFormat2018(data)
}
\arguments{
\item{data}{Dataframe containing SingSparrow data}
}
\description{
Apply the 2018 format to database
}
\keyword{SingSparrow,}
\keyword{conditioning,}
\keyword{format}
\keyword{operant}
