% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByLastPb.R
\name{filterByLastPb}
\alias{filterByLastPb}
\title{Filter SingSparrow OC according to last playbacks}
\usage{
filterByLastPb(data, filtype = NA, ...)
}
\arguments{
\item{data}{Dataframe containing SingSparrow data.}

\item{filtype}{Type of filter. Can the take the values 'onesong',
'bothsongs', or 'bothsongs_plus100'.}

\item{...}{Other arguments passed to getLastPb().}
}
\description{
Filter SingSparrow OC according to last playbacks
}
\details{
This function filters SingSparrow data according to the last time a
  song was played back. The type of filter is specified via the argument
  filtype. A value of 'onesong' means that the dataframe will be filtered to
  contain all records per day until playback of one of the songs was
  exhausted. A value of 'bothsongs', until playback of both songs were
  exahusted. A value of 'bothsongs_plus100', 100 presses after playback of
  both songs were exhausted.
}
\keyword{SingSparrow}
\keyword{conditioning,}
\keyword{operant}
