% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAge.R
\name{getAge}
\alias{getAge}
\title{Get age of a bird}
\usage{
getAge(bird, dates, birdinfo)
}
\arguments{
\item{bird}{The id of the bird.}

\item{dates}{Dates for which the age of the bird is wished to be known.}

\item{birdinfo}{Dataframe contaning hatching records of the bird.}
}
\description{
Get the age of a bird for any date.
}
\keyword{age,}
\keyword{bird}
