% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKeyCount.R
\name{getKeyCount}
\alias{getKeyCount}
\title{Get daily count of presses for each key.}
\usage{
getKeyCount(data)
}
\arguments{
\item{data}{Data frame containing output data from SingSparrow for a single
bird}

\item{excludeRev}{Logical. Exlude data from some days after a contingency
reversal.}

\item{nexlude}{If excludeRev is TRUE, the number of days after a contingency
reversal to exclude.}
}
\description{
Get daily count of presses for each key.
}
\details{
All the measures based on presses for a key associated with playback
  of a given song.
}
\keyword{SingSparrow}
\keyword{conditioning,}
\keyword{preference,operant}
\keyword{songbirds,}
