% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastPb.R
\name{getLastPb}
\alias{getLastPb}
\title{Get record indexes when playback of a song or more were exhausted}
\usage{
getLastPb(data, quota = 30)
}
\arguments{
\item{data}{Dataframe containing SingSparrow data.}

\item{quota}{Quota for daily playback of songs.}
}
\description{
Get record indexes when playback of a song or more were exhausted
}
\keyword{conditioning,}
\keyword{exhausted.}
\keyword{operant}
\keyword{quota}
