% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPreference.R
\name{getPreference}
\alias{getPreference}
\title{Estimate preference for a song}
\usage{
getPreference()
}
\arguments{
\item{data}{Data frame containing output data from SingSparrow}

\item{song}{Song for which preference is measured.}

\item{type}{Type of preference measurement. See details.}

\item{excludeRev}{Logical. Exlude data from some days after a contingency
reversal.}

\item{nexlude}{If excludeRev is TRUE, the number of days after a contingency
reversal to exclude.}
}
\description{
Estimate preference for a song
}
\details{
All the measures based on presses for a key
  associated with playback of a given song.
}
\keyword{SingSparrow}
\keyword{conditioning,}
\keyword{preference,operant}
\keyword{songbirds,}
