% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBirdinfo.R
\name{importBirdinfo}
\alias{importBirdinfo}
\title{Import information on birds}
\usage{
importBirdinfo(conn)
}
\arguments{
\item{conn}{Connection from which to import the data.}
}
\description{
Import information on birds
}
\keyword{birds}
\keyword{database,}
