% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOC.R
\name{loadOC}
\alias{loadOC}
\title{Load output files from SingSparrow}
\usage{
loadOC(birDir, bird = NULL, datalim = NA, exclude = NA,
  fmt = "2015", zerokeyrm = T)
}
\arguments{
\item{birDir}{Folder containing output files.}

\item{bird}{Id of the bird.}

\item{datalim}{Number of records from each day to be kept}

\item{fmt}{Format of keypress data file.}

\item{zerokeyrm}{Logical. Should rows with values of zero for key be removed?}

\item{exlude}{Indexes of records to exclude.}
}
\value{
A database containing values from SingSparrow! output in
  human-readable format.
}
\description{
Load output files from SingSparrow
}
\details{
If bird is not provided, then it is assumed to be the basename of
  birDir.

  The exclude indexes apply to the dataframe output by the plotting function.
  The indexes should be retrieved after preliminary exploration of the
  database.

  Possible formats of the keypress data file are "2015" or "2019".
}
\keyword{conditioning,singsparrow}
\keyword{operant}
