% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmRev.R
\name{rmRev}
\alias{rmRev}
\title{Remove days after a reversal}
\usage{
rmRev(keypress, bird, birdinfo, dpostrev, verbose = F)
}
\arguments{
\item{keypress}{Dataframe containing keypress data from SingSparrow}

\item{bird}{Name of the bird.}

\item{birdinfo}{Optional. Database on which to check accuracy of reversal
record}

\item{dpostrev}{Number of days after a reversal that data will be excluded}

\item{verbose}{Logical. Should verbose mode be activated?}
}
\value{
Dataframe containing info on keypresses, in same format as in
  keypress, excluding data some days after a reversal was performed.
}
\description{
Remove data from a certain number of days after a contingency reversal. The
reversal must be recorded on a database containing information on
experimental treatment applied to the bird.
}
\keyword{conditioning}
\keyword{operant}
\keyword{reversal,}
