% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekRev.R
\name{seekRev}
\alias{seekRev}
\title{Search contingency reversals in dataframe}
\usage{
seekRev(x)
}
\arguments{
\item{x}{Dataframe containing records from SingSparrow.}
}
\value{
A dataframe with the reversal time. This time is the midpoint between
  two consecutive events found in x before and after the reversal. In
  addition, the dataframe contains the song which became associated withn key
  A after the reversal, and that which became associated with key B.
}
\description{
Search contingency reversals in dataframe
}
\keyword{conditioning}
\keyword{operant}
\keyword{reversal,}
