% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shouldISwitch.R
\name{shouldISwitch}
\alias{shouldISwitch}
\title{Take the decission  of whether to set-up a contingency reversal based on the
history of key presses.}
\usage{
shouldISwitch(data, alpha = 0.05, consecutive_days = 5, ...)
}
\arguments{
\item{data}{Dataframe to analyze.}

\item{alpha}{Alpha level for stablishing a significant preference.}

\item{consecutive_days}{Number of consecutive days when preference for a song
must be significant.}

\item{...}{Arguments passed to filterByLastPb.}
}
\description{
Take the decission  of whether to set-up a contingency reversal based on the
history of key presses.
}
\details{
the function extracts the last `consecutive_days` rows of `data` and
  for each day runs a permutation test to test whether the number of presses
  for one key was significantly larger than for the other. If the number of
  presses was significantly higher for one of the keys during the
  `consecutive_days`, then a reversal is recommended.

  In some reinforcement schedules, reversals are recommended only once. Check
  with your reinforcement schedule to do a reversal before accepting the
  recommendation given by this function.
}
\keyword{conditioning,}
\keyword{operant}
\keyword{reversal}
