% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOC.R
\name{loadOC}
\alias{loadOC}
\title{Load output files from SingSparrow! for a given bird}
\usage{
loadOC(bird = NA, birDir = "../data/", datalim = NA, exclude = NA,
  col_names = c("Event", "Key", "Sound", "File", "Year", "Month", "Day",
  "Hour", "Min", "Sec"))
}
\arguments{
\item{bird}{Id of the bird.}

\item{birDir}{Folder containing output files.}

\item{datalim}{Number of records from each day to be plotted.}

\item{col_names}{Names of the columns in SingSparrow output.}

\item{exlude}{Indexes of records to exclude from plot.}
}
\value{
A database containing values from SingSparrow! output in
  human-readable format.
}
\description{
Load output files from SingSparrow! for a given bird
}
\details{
If bird is not provided, then it is assumed to be the basename of
  birDir.

  The exclude indexes apply to the dataframe output by the plotting
  function. The indexes should be retrieved after preliminary exploration of
  the database.
}
\keyword{conditioning,singsparrow}
\keyword{operant}
